# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Archaeological Excavation Management System API",
    "description": "Backend API for 3D stratigraphy, offline PWA sync, QR chain-of-custody, and ML material classification.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/sites": {
      "get": {
        "tags": [
          "Sites"
        ],
        "summary": "Get Sites",
        "operationId": "get_sites_api_v1_sites_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "historical_period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Historical Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SiteResponse"
                  },
                  "title": "Response Get Sites Api V1 Sites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sites"
        ],
        "summary": "Create Site",
        "operationId": "create_site_api_v1_sites_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sites/{site_id}": {
      "get": {
        "tags": [
          "Sites"
        ],
        "summary": "Get Site",
        "operationId": "get_site_api_v1_sites__site_id__get",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Sites"
        ],
        "summary": "Update Site",
        "operationId": "update_site_api_v1_sites__site_id__patch",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sites"
        ],
        "summary": "Delete Site",
        "operationId": "delete_site_api_v1_sites__site_id__delete",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/media/upload": {
      "post": {
        "tags": [
          "Media"
        ],
        "summary": "Upload Media Record",
        "operationId": "upload_media_record_api_v1_media_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "List Media",
        "operationId": "list_media_api_v1_media_get",
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaResponse"
                  },
                  "title": "Response List Media Api V1 Media Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{media_id}": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Get Media Asset",
        "operationId": "get_media_asset_api_v1_media__media_id__get",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Media"
        ],
        "summary": "Delete Media Asset",
        "operationId": "delete_media_asset_api_v1_media__media_id__delete",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sites/{site_id}/stratigraphy": {
      "get": {
        "tags": [
          "3D Stratigraphy"
        ],
        "summary": "Get Site Stratigraphy",
        "operationId": "get_site_stratigraphy_api_v1_sites__site_id__stratigraphy_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrenchStratigraphyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "3D Stratigraphy"
        ],
        "summary": "Add Site Stratigraphic Layer",
        "operationId": "add_site_stratigraphic_layer_api_v1_sites__site_id__stratigraphy_post",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StratigraphicLayerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StratigraphicLayerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaCreate": {
        "properties": {
          "entity_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "file_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Type"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "camera_metadata": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Camera Metadata"
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "entity_id",
          "file_url"
        ],
        "title": "MediaCreate"
      },
      "MediaResponse": {
        "properties": {
          "entity_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "file_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Type"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "camera_metadata": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Camera Metadata"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "entity_id",
          "file_url",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "MediaResponse"
      },
...
      "SiteCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "site_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Site Code"
          },
          "region": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Region"
          },
          "historical_period": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Historical Period"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "altitude_meters": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude Meters"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "site_code",
          "region",
          "historical_period",
          "latitude",
          "longitude"
        ],
        "title": "SiteCreate"
      },
      "SiteResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "site_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Site Code"
          },
          "region": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Region"
          },
          "historical_period": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Historical Period"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "altitude_meters": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude Meters"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "site_code",
          "region",
          "historical_period",
          "latitude",
          "longitude",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "SiteResponse"
      },
      "SiteUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "site_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Code"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "historical_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Historical Period"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "altitude_meters": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude Meters"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "SiteUpdate"
      },
...
      "StratigraphicLayerCreate": {
        "properties": {
          "layer_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Layer Code"
          },
          "historical_period": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Historical Period"
          },
          "depth_top_meters": {
            "type": "number",
            "title": "Depth Top Meters"
          },
          "depth_bottom_meters": {
            "type": "number",
            "title": "Depth Bottom Meters"
          },
          "color_hex": {
            "type": "string",
            "maxLength": 7,
            "title": "Color Hex",
            "default": "#8B4513"
          }
        },
        "type": "object",
        "required": [
          "layer_code",
          "historical_period",
          "depth_top_meters",
          "depth_bottom_meters"
        ],
        "title": "StratigraphicLayerCreate"
      },
      "StratigraphicLayerResponse": {
        "properties": {
          "layer_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Layer Code"
          },
          "historical_period": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Historical Period"
          },
          "depth_top_meters": {
            "type": "number",
            "title": "Depth Top Meters"
          },
          "depth_bottom_meters": {
            "type": "number",
            "title": "Depth Bottom Meters"
          },
          "color_hex": {
            "type": "string",
            "maxLength": 7,
            "title": "Color Hex",
            "default": "#8B4513"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "layer_code",
          "historical_period",
          "depth_top_meters",
          "depth_bottom_meters",
          "id",
          "site_id",
          "created_at",
          "updated_at"
        ],
        "title": "StratigraphicLayerResponse"
      },
...
      "TrenchBounds": {
        "properties": {
          "min_x": {
            "type": "number",
            "title": "Min X"
          },
          "max_x": {
            "type": "number",
            "title": "Max X"
          },
          "min_y": {
            "type": "number",
            "title": "Min Y"
          },
          "max_y": {
            "type": "number",
            "title": "Max Y"
          },
          "min_depth": {
            "type": "number",
            "title": "Min Depth"
          },
          "max_depth": {
            "type": "number",
            "title": "Max Depth"
          }
        },
        "type": "object",
        "required": [
          "min_x",
          "max_x",
          "min_y",
          "max_y",
          "min_depth",
          "max_depth"
        ],
        "title": "TrenchBounds"
      },
      "TrenchStratigraphyResponse": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "bounds": {
            "$ref": "#/components/schemas/TrenchBounds"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/StratigraphicLayerResponse"
            },
            "type": "array",
            "title": "Layers"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/SpatialArtifactNode"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "site_name",
          "bounds",
          "layers",
          "artifacts"
        ],
        "title": "TrenchStratigraphyResponse"
      },
...
    }
  }
}